+++ noodles-sam/src/reader/record/quality_scores.rs
@@ -7,6 +7,13 @@ use crate::record::{quality_scores, QualityScores};
 pub enum ParseError {
     /// The input is empty.
     Empty,
+    /// The length does not match the sequence length.
+    LengthMismatch {
+        /// The actual length.
+        actual: usize,
+        /// The expected length.
+        expected: usize,
+    },
     /// The input is invalid.
     Invalid(quality_scores::ParseError),
 }
@@ -14,8 +21,8 @@ pub enum ParseError {
 impl error::Error for ParseError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
-            Self::Empty => None,
             Self::Invalid(e) => Some(e),
+            _ => None,
         }
     }
 }
@@ -24,6 +31,9 @@ impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Empty => write!(f, "empty input"),
+            Self::LengthMismatch { actual, expected } => {
+                write!(f, "length mismatch: expeced {expected}, got {actual}")
+            }
             Self::Invalid(_) => write!(f, "invalid input"),
         }
     }
@@ -31,12 +41,18 @@ impl fmt::Display for ParseError {
 
 pub(crate) fn parse_quality_scores(
     src: &[u8],
+    sequence_len: usize,
     quality_scores: &mut QualityScores,
 ) -> Result<(), ParseError> {
     const OFFSET: u8 = b'!';
 
     if src.is_empty() {
         return Err(ParseError::Empty);
+    } else if src.len() != sequence_len {
+        return Err(ParseError::LengthMismatch {
+            actual: src.len(),
+            expected: sequence_len,
+        });
     }
 
     let raw_quality_scores = Vec::from(mem::take(quality_scores));
@@ -58,19 +74,28 @@ mod tests {
         let mut quality_scores = QualityScores::default();
 
         quality_scores.clear();
-        parse_quality_scores(b"NDLS", &mut quality_scores)?;
+        parse_quality_scores(b"NDLS", 4, &mut quality_scores)?;
         let expected = QualityScores::try_from(vec![45, 35, 43, 50])?;
         assert_eq!(quality_scores, expected);
 
         quality_scores.clear();
         assert_eq!(
-            parse_quality_scores(b"", &mut quality_scores),
+            parse_quality_scores(b"", 0, &mut quality_scores),
             Err(ParseError::Empty)
         );
 
         quality_scores.clear();
+        assert_eq!(
+            parse_quality_scores(b"NDLS", 2, &mut quality_scores),
+            Err(ParseError::LengthMismatch {
+                actual: 4,
+                expected: 2
+            })
+        );
+
+        quality_scores.clear();
         assert!(matches!(
-            parse_quality_scores(&[0x07], &mut quality_scores),
+            parse_quality_scores(&[0x07], 1, &mut quality_scores),
             Err(ParseError::Invalid(_))
         ));