+++ noodles-sam/CHANGELOG.md
@@ -34,6 +34,9 @@
 
     If present, the quality scores must have at least one score.
 
+  * sam/reader/record/quality_scores: Ensure quality scores length matches
+    sequence length.
+
   * sam/record/data: `Data::get`, `Data::get_index_of`, and `Data::remove` take
     a reference of an equivalent tag.